app/testpmd: fix RSS flags size
Since commit 8a387fa85f02 ("ethdev: more RSS flags") in DPDK 1.7,
RSS flags have increased.
According to rss_hf definition in rte_eth_rss_conf, it shall be uint64 type.
Using uint16 will get truncated value, and cause incorrect output. This
fix corrected this issue.

Acked-by: Thomas Monjalon <thomas.monjalon@6wind.com>

+++ app/test-pmd/config.c
@@ -824,7 +824,7 @@ port_rss_hash_conf_show(portid_t port_id, int show_rss_key)
 {
 	struct rte_eth_rss_conf rss_conf;
 	uint8_t rss_key[10 * 4];
-	uint16_t rss_hf;
+	uint64_t rss_hf;
 	uint8_t i;
 	int diag;
